Método de cadastro para Student e SchoolReport implementados Além da implementação das classes Student (estudante) e SchoolReport (boletim), também foram feitas pequenas correções no tipo de target dos métodos de busca e foi implementada a regra de negócios do cadastro. Sendo possível, agora, escolher entre o cadastro de um aluno e a inserção de um boletim. Quando o cadastro de um aluno (estudante) é criado uma nova instância da classe Student é criada. Quando um novo boletim é criado, a sua instância é adicionada a classe Student, pois a classe Student contêm um SchoolReport (por composição).

+++ src/ricardo_paulo/net/Search.java
@@ -2,7 +2,7 @@ package ricardo_paulo.net;
 
 public class Search {
 
-    public static Student sequentialSearch (Student[] students, int target) {
+    public static Student sequentialSearch (Student[] students, long target) {
 
         for (int i = 0; i < students.length; i++) {
             if (students[i].getRegistry() == target)
@@ -12,7 +12,7 @@ public class Search {
         return new Student();
     }
 
-    public static Student binarySearch (Student[] students, int target) {
+    public static Student binarySearch (Student[] students, long target) {
         int start = 0;
         int finish = students.length - 1;
